move test helper functions to files manager
